Temporary add a patch to clear tripleo promotion
vitrage_tempest_plugin is being refactored and is not
working and breaking tripleo jobs, as a quick fix Temporary
applying patches in spec, it will be reverted as soon as
vitrage_tempest_plugin is fixed.
Such issues should also be handled in TripleO(vitrage is not tested
in TripleO but it's affecting the jobs).

Related-Bug: lp#1817154
Change-Id: I31f97afc3b6c5f6a26dafa90780918b043040e0d